Add mechanism to define static hostnames in the hostResolver
This is an alternative to adding them globally to /etc/hosts on the
host (not the guest). It also allows aliasing them to host.lima.internal.

This allows names to be resolved not just inside the guest, but also inside
containers inside the guest, which only have access to /etc/resolv.conf, and
not the full resolver inside the guest.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>