iptables rule without an ip range applies to all interfaces
The current logic pretends it only applies to 127.0.0.1, which
means the new `guestIPMustBeZero` rule does not detect it properly.

This is a problem with nerdctl and containerd on Alpine. It works
"by accident" on Ubuntu because the port was also bound to [::],
which had an entry in /proc/net/tcp6.

Signed-off-by: Jan Dubois <jan.dubois@suse.com>