Return NODATA for IPv6 AAAA question types
Previously we were replying with A response for AAAA
queries which was causing an error in some clients
e.g nginx. The new approach waits for 1 second and then
replies with NODATA. This approach allows for more semantic
reply even when client uses same Transaction ID for both
A and AAAA questions.

Signed-off-by: Nino Kodabande <nkodabande@suse.com>