docs: Usage: remove scaring caution
`Lima may have bugs that result in loss of data` sounds too scary for
the current version of Lima.

It is still true that Lima, as well as any software, may have such bugs,
but it does no longer need to be particularly documented here.

In Lima v0.14.0, we had a regression (issue 1242, fixed in v0.14.1) that
accidentally leaved `/var/lib/docker` in Colima instances unmounted,
but the actualy data were not lost, and could be easily recovered by
downgrading or upgrading Lima to other versions.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>