qemu, aarch64: use EDK2 from "Firmware/edk2 20231213 patches"
The "Firmware/edk2 20231213" patches
<https://lists.gnu.org/archive/html/qemu-devel/2023-12/msg01694.html>
are necessary to boot most aarch64 images (except Debian, which does not use
UEFI shim).

The patches are proposed for the QEMU v8.2.0 milestone, but likely to be
postponed to v8.2.1.
Until the patches get accepted in the QEMU upstream, Lima fetches the patched
edk2 binary from
<https://gitlab.com/kraxel/qemu/-/tags/firmware%2Fedk2-20231213-pull-request>.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>